Adds a nice blend of performance
