+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "axios": "^1.7.9",
+    "critters": "^0.0.23",
     "next": "^15.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
